refactor: move formatRelativeTime to utils/display.ts
